+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2022-07-11
+
+### Added
+
+- Ability to use `RSET` SMTP command, following [RFC 821](https://www.rfc-editor.org/rfc/rfc821)
+- Ability to configure multiple message receiving flow during one session. Thanks [@Pointer666](https://github.com/Pointer666) for report.
+
+### Updated
+
+- Updated `server`, tests
+- Updated `message`, tests
+- Updated `configuration`, `configurationAttr`, tests
+- Updated command handlers, tests
+- Updated cmd, consts
+- Updated package documentation, changelog
+
+### Removed
+
+- Removed `message#cleared`, `message#isCleared?()`
+
 ## [1.8.1] - 2022-05-26
 
 ### Fixed